moved a test class to the actual test folder. and there was an issue where Employee still showed up in it's package, while it was removed from it's package and moved to the main folder models. not sure if this weird issue is now fixed everywhere, so keep that in mind.
